chore(material): move material module into modules folder
