+++ src/client/src/app/app.module.ts
@@ -26,7 +26,7 @@ import { FaModule } from './modules/fa/fa.module';
 import { MerchComponent } from './components/merch/merch.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { LayoutModule } from '@angular/cdk/layout';
-import { MaterialModule } from './material/material.module';
+import { MaterialModule } from './modules/material/material.module';
 import { MerchDetailsComponent } from './components/merch-details/merch-details.component';
 import { MainNavComponent } from './main-nav/main-nav.component';
 import { StoreRouterConnectingModule } from '@ngrx/router-store';
